Trait ic_consensus_message::ConsensusMessageHashable

pub trait ConsensusMessageHashable: Clone {
    fn get_id(&self) -> ConsensusMessageId;

    fn get_cm_hash(&self) -> ConsensusMessageHash;

    fn assert(msg: &ConsensusMessage) -> Option<&Self>;

    fn to_message(self) -> ConsensusMessage;

    fn check_integrity(&self) -> bool { ... }
}

Required methods

fn get_id(&self) -> ConsensusMessageId

fn get_cm_hash(&self) -> ConsensusMessageHash

fn assert(msg: &ConsensusMessage) -> Option<&Self>

fn to_message(self) -> ConsensusMessage

Loading content...
Provided methods

fn check_integrity(&self) -> bool

Check integrity of a message. Default is true. This should be implemented for those that have Hashed<H, V>. Note that if lazy loading is also used, it will force evaluation.

Loading content...